# models/login_success_response_definition.go
// Code generated by go-swagger; DO NOT EDIT.

package models

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	"github.com/go-openapi/strfmt"
	"github.com/go-openapi/swag"
)

// LoginSuccessResponseDefinition login success response definition
//
// swagger:model LoginSuccessResponseDefinition
type LoginSuccessResponseDefinition struct {

	// token
	Token string `json:"token,omitempty"`
}

// Validate validates this login success response definition
func (m *LoginSuccessResponseDefinition) Validate(formats strfmt.Registry) error {
	return nil
}

// MarshalBinary interface implementation
func (m *LoginSuccessResponseDefinition) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *LoginSuccessResponseDefinition) UnmarshalBinary(b []byte) error {
	var res LoginSuccessResponseDefinition
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}