# models/update_users_definition.go
// Code generated by go-swagger; DO NOT EDIT.

package models

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	"github.com/go-openapi/errors"
	"github.com/go-openapi/strfmt"
	"github.com/go-openapi/swag"
	"github.com/go-openapi/validate"
)

// UpdateUsersDefinition update users definition
//
// swagger:model UpdateUsersDefinition
type UpdateUsersDefinition struct {

	// email
	// Required: true
	Email *string `json:"Email"`

	// jwt token
	// Required: true
	JwtToken *string `json:"JwtToken"`

	// name
	// Required: true
	Name *string `json:"Name"`

	// password
	// Required: true
	Password *string `json:"Password"`
}

// Validate validates this update users definition
func (m *UpdateUsersDefinition) Validate(formats strfmt.Registry) error {
	var res []error

	if err := m.validateEmail(formats); err != nil {
		res = append(res, err)
	}

	if err := m.validateJwtToken(formats); err != nil {
		res = append(res, err)
	}

	if err := m.validateName(formats); err != nil {
		res = append(res, err)
	}

	if err := m.validatePassword(formats); err != nil {
		res = append(res, err)
	}

	if len(res) > 0 {
		return errors.CompositeValidationError(res...)
	}
	return nil
}

func (m *UpdateUsersDefinition) validateEmail(formats strfmt.Registry) error {

	if err := validate.Required("Email", "body", m.Email); err != nil {
		return err
	}

	return nil
}

func (m *UpdateUsersDefinition) validateJwtToken(formats strfmt.Registry) error {

	if err := validate.Required("JwtToken", "body", m.JwtToken); err != nil {
		return err
	}

	return nil
}

func (m *UpdateUsersDefinition) validateName(formats strfmt.Registry) error {

	if err := validate.Required("Name", "body", m.Name); err != nil {
		return err
	}

	return nil
}

func (m *UpdateUsersDefinition) validatePassword(formats strfmt.Registry) error {

	if err := validate.Required("Password", "body", m.Password); err != nil {
		return err
	}

	return nil
}

// MarshalBinary interface implementation
func (m *UpdateUsersDefinition) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *UpdateUsersDefinition) UnmarshalBinary(b []byte) error {
	var res UpdateUsersDefinition
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}